refactor: type 도메인별 분리
