+++ src/types/error.type.ts
@@ -1,0 +1,6 @@
+interface ErrorState {
+  is: boolean;
+  message: string;
+}
+
+export type { ErrorState };
